functions/requirements.txt
cachetools==3.1.1
certifi==2019.9.11
google-api-core==1.26.3
google-auth==1.28.1
google-cloud-bigquery==2.13.1
google-cloud==0.34.0
google-cloud-pubsub==2.4.1
googleapis-common-protos==1.53.0
pandas==1.2.3
requests==2.22.0
pyarrow==3.0.0
urllib3==1.25.6